fix line-feeds, set svn:eol-style
git-svn-id: https://svn.apache.org/repos/asf/ant/core/trunk@279394 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/org/apache/tools/ant/taskdefs/optional/ssh/AbstractSshMessage.java
@@ -157,7 +157,7 @@ public abstract class AbstractSshMessage {
      */
     protected void logStats(long timeStarted,
                              long timeEnded,
-                             int totalLength) {
+                             long totalLength) {
         double duration = (timeEnded - timeStarted) / 1000.0;
         NumberFormat format = NumberFormat.getNumberInstance();
         format.setMaximumFractionDigits(2);
@@ -184,7 +184,7 @@ public abstract class AbstractSshMessage {
      * @param percentTransmitted the current percent transmitted
      * @return the percent that the file is of the total
      */
-    protected final int trackProgress(int filesize, int totalLength,
+    protected final int trackProgress(long filesize, long totalLength,
                                       int percentTransmitted) {
 
         int percent = (int) Math.round(Math.floor((totalLength
